Cambio para la manipulacion de inventario, proveedores con usuario y manejo segun estado

+++ src/main/java/com/sanisidro/restaurante/features/products/dto/inventory/response/InventoryResponse.java
@@ -2,6 +2,7 @@ package com.sanisidro.restaurante.features.products.dto.inventory.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -15,8 +16,8 @@ public class InventoryResponse {
     private String ingredientName;
     private String unitName;
     private String unitSymbol;
-    private Integer currentStock;
-    private Integer minimumStock;
+    private BigDecimal currentStock;
+    private BigDecimal minimumStock;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
