+++ src/main/java/com/sanisidro/restaurante/features/products/dto/inventorymovement/request/InventoryMovementRequest.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -23,7 +24,7 @@ public class InventoryMovementRequest {
 
     @NotNull(message = "La cantidad es obligatoria")
     @Min(value = 1, message = "La cantidad debe ser mayor a 0")
-    private Integer quantity;
+    private BigDecimal quantity;
 
     private String reason;
