+++ src/main/java/com/sanisidro/restaurante/features/products/dto/inventorymovement/response/InventoryMovementResponse.java
@@ -3,6 +3,8 @@ package com.sanisidro.restaurante.features.products.dto.inventorymovement.respon
 import com.sanisidro.restaurante.features.orders.enums.MovementSource;
 import com.sanisidro.restaurante.features.products.enums.MovementType;
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -17,7 +19,7 @@ public class InventoryMovementResponse {
     private String unitName;
     private String unitSymbol;
     private MovementType type;
-    private Integer quantity;
+    private BigDecimal quantity;
     private LocalDateTime date;
     private String reason;
     private MovementSource source;
